Moving the ActionTypes.ts location and renaming it.
